+++ src/headless/plugins/muc/muc.js
@@ -845,6 +845,12 @@ const ChatRoomMixin = {
 
     async close (ev) {
         await this.leave();
+        if (
+            api.settings.get('auto_register_muc_nickname') === 'unregister' &&
+            (await api.disco.supports(Strophe.NS.MUC_REGISTER, this.get('jid')))
+        ) {
+            this.unregisterNickname();
+        }
         this.occupants.clearStore();
 
         if (ev?.name !== 'closeAllChatBoxes' && api.settings.get('muc_clear_messages_on_leave')) {
@@ -1533,7 +1539,7 @@ const ChatRoomMixin = {
 
     async registerNickname () {
         // See https://xmpp.org/extensions/xep-0045.html#register
-        const __ = _converse.__;
+        const { __ } = _converse;
         const nick = this.get('nick');
         const jid = this.get('jid');
         let iq, err_msg;
@@ -1541,7 +1547,6 @@ const ChatRoomMixin = {
             iq = await api.sendIQ(
                 $iq({
                     'to': jid,
-                    'from': _converse.connection.jid,
                     'type': 'get'
                 }).c('query', { 'xmlns': Strophe.NS.MUC_REGISTER })
             );
@@ -1562,19 +1567,13 @@ const ChatRoomMixin = {
             await api.sendIQ(
                 $iq({
                     'to': jid,
-                    'from': _converse.connection.jid,
                     'type': 'set'
-                })
-                    .c('query', { 'xmlns': Strophe.NS.MUC_REGISTER })
+                }).c('query', { 'xmlns': Strophe.NS.MUC_REGISTER })
                     .c('x', { 'xmlns': Strophe.NS.XFORM, 'type': 'submit' })
-                    .c('field', { 'var': 'FORM_TYPE' })
-                    .c('value')
-                    .t('http://jabber.org/protocol/muc#register')
-                    .up()
-                    .up()
-                    .c('field', { 'var': 'muc#register_roomnick' })
-                    .c('value')
-                    .t(nick)
+                        .c('field', { 'var': 'FORM_TYPE' })
+                            .c('value').t('http://jabber.org/protocol/muc#register').up().up()
+                        .c('field', { 'var': 'muc#register_roomnick' })
+                            .c('value').t(nick)
             );
         } catch (e) {
             if (sizzle(`service-unavailable[xmlns="${Strophe.NS.STANZAS}"]`, e).length) {
@@ -1585,6 +1584,28 @@ const ChatRoomMixin = {
             log.error(err_msg);
             log.error(e);
             return err_msg;
+        }
+    },
+
+    async unregisterNickname () {
+        const jid = this.get('jid');
+        let iq;
+        try {
+            iq = await api.sendIQ(
+                $iq({
+                    'to': jid,
+                    'type': 'set'
+                }).c('query', { 'xmlns': Strophe.NS.MUC_REGISTER })
+            );
+        } catch (e) {
+            log.error(e);
+            return e;
+        }
+        if (sizzle(`query[xmlns="${Strophe.NS.MUC_REGISTER}"] registered`, iq).pop()) {
+            const iq = $iq({ 'to': jid, 'type': 'set' })
+                .c('query', { 'xmlns': Strophe.NS.MUC_REGISTER })
+                .c('remove');
+            return api.sendIQ(iq).catch(e => log.error(e));
         }
     },